restructures dirs.
