+++ README.md
@@ -1,7 +1,2 @@
 # erfgoedbot
 Facebook Messenger bot with Dutch erfgoed data, linked to Wikidata
-
-deploy naar heroku:
-```sh
-git subtree push --prefix src heroku master
-```
