+++ etc/message.json
@@ -1,0 +1,25 @@
+{
+    "entry": [
+        {
+            "id": 1397278930293852,
+            "messaging": [
+                {
+                    "message": {
+                        "mid": "mid.1460620432888:f8e3412003d2d1cd93",
+                        "seq": 12604,
+                        "text": ":MSG:"
+                    },
+                    "recipient": {
+                        "id": 1397278930293852
+                    },
+                    "sender": {
+                        "id": 1397278930293852
+                    },
+                    "timestamp": 1460620433123
+                }
+            ],
+            "time": 1460620433256
+        }
+    ],
+    "object": "page"
+}
