+++ src/bot/search.js
