+++ src/package.json
@@ -16,8 +16,6 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.15.0",
-    "config": "^1.20.4",
-    "ejs": "^2.4.2",
     "express": "^4.13.4",
     "lodash": "^4.16.6",
     "request": "^2.72.0",
